Don't attach gesture to chart when time can't be set
Attaching the gesture means it will always be reacting to inputs, and
may cause other gestures in the same area to not be called. Adding a
disabled modifier may produce the same effect, but it seemed too high
risk to add that now, and simply not attaching the gesture works fine
too.